+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
+## [Unreleased]
+#### Fixed
+- A bug when the user can not enter a value in input-field with type 'number' on Firefox
+
 ## [1.10.3-rc.0] - 2019-10-11
 #### Added
 - Integration with Erc20 and stellar